requirements.txt
cffi==1.9.1
Pillow==4.0.0
